+++ Changes.md
@@ -7,7 +7,7 @@ Refactor the underlying implementation to prepare for further development
 New features
 
 - Add `arrow=53` support
-- Add `Binary` / `LargeBinary` support for `arrow2`
+- Add `Binary`, `LargeBinary`, `FixedSizeBinary(n)`, `FixedSizeList(n)` support for `arrow2`
 - Add support to serialize / deserialize `bool` from integer arrays
 - Add a helper to construct `Bool8` arrays
 - Include the path of the field that caused an error in the error message